requirements.txt
keras==2.15.0
matplotlib==3.7.2
numpy==1.24.3
Pillow==9.4.0
Pillow==10.1.0
scikit_learn==1.3.0
seaborn==0.13.0
streamlit==1.28.2
tensorflow==2.15.0
tensorflow_intel==2.15.0
